move BaseModelWorker outside serve.model_worker to make it independent (#2531)
